+++ src/EventDispatcher/EventDispatcherTrait.php
@@ -1,4 +1,5 @@
 <?php
+declare(strict_types = 1);
 
 namespace Skautis\EventDispatcher;
 
@@ -9,35 +10,26 @@ trait EventDispatcherTrait
     private $listeners = [];
 
 
-    /**
-     * @param string|null $eventName
-     * @return bool
-     */
-    protected function hasListeners($eventName = null)
+    protected function hasListeners(?string $eventName = null): bool
     {
         return $eventName === null ? !empty($this->listeners) : !empty($this->listeners[$eventName]);
     }
 
     /**
-     * @param string $eventName
      * @param mixed $data
      */
-    protected function dispatch($eventName, $data)
+    protected function dispatch(string $eventName, $data): void
     {
         if (!$this->hasListeners($eventName)) {
             return;
         }
 
         foreach ($this->listeners[$eventName] as $callback) {
-            call_user_func($callback, $data);
+            $callback($data);
         }
     }
 
-    /**
-     * @param string $eventName
-     * @param callable $callback
-     */
-    public function subscribe($eventName, callable $callback)
+    public function subscribe(string $eventName, callable $callback): void
     {
         $this->listeners[$eventName][] = $callback;
     }