+++ src/SessionAdapter/AdapterInterface.php
@@ -1,4 +1,5 @@
 <?php
+declare(strict_types = 1);
 
 namespace Skautis\SessionAdapter;
 
@@ -10,22 +11,18 @@ interface AdapterInterface
 
     /**
      * Ulozi data do session
-     *
-     * @return void
      */
-    public function set($name, $object);
+    public function set(string $name, $object): void ;
 
     /**
      * Overi existenci dat v session
-     *
-     * @return bool
      */
-    public function has($name);
+    public function has(string $name): bool;
 
     /**
      * Ziska data ze session
      *
      * @return mixed
      */
-    public function get($name);
+    public function get(string $name);
 }