+++ src/StaticClassException.php
@@ -1,12 +1,15 @@
 <?php
+declare(strict_types = 1);
 
 namespace Skautis;
+
+use LogicException;
 
 /**
  * Vyhozena v případě pokusu o vytvoření instance statické třídy.
  *
  * @author Petr Morávek <petr@pada.cz>
  */
-class StaticClassException extends \LogicException implements Exception
+class StaticClassException extends LogicException implements Exception
 {
 }