+++ src/Wsdl/WebServiceFactory.php
@@ -1,6 +1,9 @@
 <?php
+declare(strict_types = 1);
 
 namespace Skautis\Wsdl;
+
+use Skautis\InvalidArgumentException;
 
 /**
  * @inheritdoc
@@ -12,18 +15,19 @@ class WebServiceFactory implements WebServiceFactoryInterface
     protected $class;
 
 
-    /**
-     * @param string $class
-     */
-    public function __construct($class = '\Skautis\Wsdl\WebService')
+    public function __construct(string $className = WebService::class)
     {
-        $this->class = $class;
+       if (!is_a($className, WebServiceInterface::class, true)) {
+         throw new InvalidArgumentException("Argument must be class name of a class implementing WebServiceInterface. '$className' given");
+       }
+
+        $this->class = $className;
     }
 
     /**
      * @inheritdoc
      */
-    public function createWebService($url, array $options)
+    public function createWebService(string $url, array $options): WebServiceInterface
     {
         return new $this->class($url, $options);
     }