+++ src/Wsdl/WebServiceFactoryInterface.php
@@ -1,4 +1,5 @@
 <?php
+declare(strict_types = 1);
 
 namespace Skautis\Wsdl;
 
@@ -13,7 +14,7 @@ interface WebServiceFactoryInterface
      *
      * @param string $url Adresa WSDL souboru
      * @param array $options Globální nastavení pro všechny požadavky
-     * @return mixed
+     * @return WebServiceInterface
      */
-    public function createWebService($url, array $options);
+    public function createWebService(string $url, array $options): WebServiceInterface;
 }