+++ CHANGELOG.md
@@ -7,6 +7,9 @@
 ### v3.0.0
 * Požadována verze PHP >=7.1
 * Odstranění podpory pro HHVM (HHVM dále [nedodržuje kompabilitu s PHP](https://hhvm.com/blog/2018/09/12/end-of-php-support-future-of-hack.html))
+* Scalar typehints pro metody
+* [Strict types](http://php.net/manual/en/functions.arguments.php#functions.arguments.type-declaration.strict) - zpětně nekompatibilní
+* Event dispatcher používá string místo int pro ``event name`` - zpětně nekompatibilní
 
 ## Verze 2.x